fixed bug in metronome (the new one was int16), and in make ramps (which didn't work with stereo)

+++ combio/_helpers.py
@@ -2,8 +2,6 @@ import os
 import importlib.resources as pkg_resources
 import warnings
 
-import matplotlib.axes
-
 from combio._warnings import framerounding
 
 import scipy.signal
@@ -131,7 +129,7 @@ def get_sound_with_metronome(samples: np.ndarray,
         metronome_samples = resampled
 
     # change amplitude if necessary
-    metronome_samples *= metronome_amplitude
+    metronome_samples *= np.float64(metronome_amplitude)
 
     for onset in onsets:
         start_pos = int(onset * fs / 1000)
